feat(overmind-vue): drop Vue 2 support, make Vue 3 the default entry point
BREAKING CHANGE: Vue 2 support has been removed. The Vue 3 Composition API
hooks are now exported from the main entry point (`overmind-vue`). The
`overmind-vue/vue3` subpath still works as a backward-compatible alias.
Peer dependency is now `vue: ^3.0.0`.
