+++ packages/overmind-vue/src/index.ts
@@ -1,177 +1,201 @@
-import { ENVIRONMENT, EventType, IContext, MODE_SSR, Overmind } from 'overmind'
-import { Component, ComponentOptions } from 'vue'
+import {
+  ENVIRONMENT,
+  EventType,
+  IConfiguration,
+  MODE_SSR,
+  Overmind,
+  IContext,
+} from 'overmind'
+import {
+  Ref,
+  inject,
+  ref,
+  onBeforeUpdate,
+  onRenderTracked,
+  onMounted,
+  onBeforeUnmount,
+  defineComponent,
+  provide,
+  h,
+  Component,
+} from 'vue'
 
-type AnyComponent = ComponentOptions | Component
-
-const OVERMIND = Symbol('OVERMIND')
 const IS_PRODUCTION = ENVIRONMENT === 'production'
 
 let nextComponentId = 0
 
-function createMixin(
-  overmind: Overmind<any>,
-  propsCallback: ((context: any) => any) | null | Function,
-  trackPropsCallback = false
-) {
+export const withOvermind = (
+  instance: Overmind<IConfiguration>,
+  Component: Component
+) => {
+  return defineComponent({
+    setup() {
+      provide('overmind', instance)
+    },
+    render() {
+      return h(Component)
+    },
+  })
+}
+
+export interface StateHook<Context extends IContext<{ state: {} }>> {
+  (): Ref<Context['state']>
+  <CB extends (state: Context['state']) => object>(
+    cb: CB
+  ): CB extends (state: Context['state']) => infer O
+    ? O extends object
+      ? Ref<O>
+      : never
+    : never
+}
+
+export function createStateHook<
+  Context extends IContext<{ state: {} }>,
+>(): StateHook<Context> {
   const componentId = nextComponentId++
   let componentInstanceId = 0
+  return ((cb: any) => {
+    const overmindInstance = inject<any>('overmind')
 
-  return {
-    beforeCreate(this: any) {
-      if ((overmind as any).mode.mode === MODE_SSR) {
-        this.overmind = {
-          state: overmind.state,
-          actions: overmind.actions,
-          effects: overmind.effects,
-          addMutationListener: overmind.addMutationListener,
-          reaction: overmind.reaction,
+    if (overmindInstance.mode.mode === MODE_SSR) {
+      return cb ? cb(overmindInstance.state) : overmindInstance.state
+    } else {
+      const overmindRef = ref<any>({})
+      const flushIds = ref(-1)
+      const { value } = overmindRef
+      const state = ref(
+        cb ? cb(overmindInstance.state) : overmindInstance.state
+      )
+
+      if (!value.tree) {
+        value.tree = overmindInstance.proxyStateTreeInstance.getTrackStateTree()
+        value.componentInstanceId = componentInstanceId++
+        value.onUpdate = (_: any, __: any, flushId: number) => {
+          value.currentFlushId = flushId
+          value.isUpdating = true
+          flushIds.value = flushId
+          state.value = {
+            ...(cb ? cb(overmindInstance.state) : overmindInstance.state),
+          }
+
+          // this.$forceUpdate()
         }
-        if (propsCallback) {
-          Object.assign(
-            this,
-            propsCallback({
-              state: overmind.state,
-              actions: overmind.actions,
-              effects: overmind.effects,
-            })
-          )
-        }
-      } else {
-        this[OVERMIND] = {
-          tree: (overmind as any).proxyStateTreeInstance.getTrackStateTree(),
-          componentInstanceId: componentInstanceId++,
-          onUpdate: (mutations: any, paths: any, flushId: any) => {
-            this[OVERMIND].currentFlushId = flushId
-            this[OVERMIND].isUpdating = true
-            this.$forceUpdate()
-          },
-          isUpdating: false,
-        }
-        this.overmind = {
-          state: this[OVERMIND].tree.state,
-          actions: overmind.actions,
-          effects: overmind.effects,
-          addMutationListener: overmind.addMutationListener,
-          reaction: overmind.reaction,
-        }
-
-        this[OVERMIND].tree.track(this[OVERMIND].onUpdate)
-
-        if (propsCallback) {
-          Object.assign(
-            this,
-            propsCallback({
-              state: this[OVERMIND].tree.state,
-              actions: overmind.actions,
-              effects: overmind.effects,
-            })
-          )
-        }
-      }
-    },
-    beforeUpdate(this: any) {
-      if ((overmind as any).mode.mode === MODE_SSR) return
-
-      this[OVERMIND].tree.track(this[OVERMIND].onUpdate)
-
-      if (propsCallback && trackPropsCallback) {
-        Object.assign(
-          this,
-          propsCallback({
-            state: this[OVERMIND].tree.state,
-            actions: overmind.actions,
-            effects: overmind.effects,
-          })
-        )
-      }
-    },
-    ...(IS_PRODUCTION
-      ? {
-          updated(this: any) {
-            this[OVERMIND].tree.stopTracking()
-          },
-        }
-      : {
-          mounted(this: any) {
-            if ((overmind as any).mode.mode === MODE_SSR) return
-
-            overmind.eventHub.emitAsync(EventType.COMPONENT_ADD, {
-              componentId,
-              componentInstanceId: this[OVERMIND].componentInstanceId,
-              name: this.$options.name || '',
-              paths: Array.from(this[OVERMIND].tree.pathDependencies) as any,
-            })
-          },
-          updated(this: any) {
-            if ((overmind as any).mode.mode === MODE_SSR) return
-
-            this[OVERMIND].tree.stopTracking()
-
-            if (this[OVERMIND].isUpdating) {
-              overmind.eventHub.emitAsync(EventType.COMPONENT_UPDATE, {
-                componentId,
-                componentInstanceId: this[OVERMIND].componentInstanceId,
-                name: this.$options.name || '',
-                flushId: this[OVERMIND].currentFlushId,
-                paths: Array.from(this[OVERMIND].tree.pathDependencies) as any,
-              })
-              this[OVERMIND].isUpdating = false
-            }
-          },
-        }),
-    beforeDestroy(this: any) {
-      if ((overmind as any).mode.mode === MODE_SSR) return
-
-      // @ts-ignore
-      overmind.proxyStateTreeInstance.disposeTree(this[OVERMIND].tree)
-      if (IS_PRODUCTION) {
-        return
+        value.isUpdating = false
       }
 
-      overmind.eventHub.emitAsync(EventType.COMPONENT_REMOVE, {
-        componentId,
-        componentInstanceId: this[OVERMIND].componentInstanceId,
-        name: this.$options.name || '',
+      onBeforeUpdate(function (this: any) {
+        if (overmindInstance.mode.mode === MODE_SSR) return
+
+        value.tree.track(value.onUpdate)
       })
-    },
+
+      onRenderTracked(function (this: any) {
+        if (IS_PRODUCTION) {
+          return
+        }
+
+        if (overmindInstance.isUpdating) {
+          overmindInstance.eventHub.emitAsync(EventType.COMPONENT_UPDATE, {
+            componentId,
+            componentInstanceId: value.componentInstanceId,
+            name: '', // this.$options.name || '',
+            flushId: value.currentFlushId,
+            paths: Array.from(value.tree.pathDependencies) as any,
+          })
+          value.isUpdating = false
+        }
+      })
+
+      onMounted(() => {
+        if (IS_PRODUCTION || overmindInstance.mode.mode === MODE_SSR) return
+        value.tree.stopTracking()
+        overmindInstance.eventHub.emitAsync(EventType.COMPONENT_ADD, {
+          componentId,
+          componentInstanceId: value.componentInstanceId,
+          name: '', // this.$options.name || '',
+          paths: Array.from(value.tree.pathDependencies) as any,
+        })
+      })
+
+      onBeforeUnmount(() => {
+        if (overmindInstance.mode.mode === MODE_SSR) return
+
+        overmindInstance.proxyStateTreeInstance.disposeTree(value.tree)
+        if (IS_PRODUCTION) {
+          return
+        }
+
+        overmindInstance.eventHub.emitAsync(EventType.COMPONENT_REMOVE, {
+          componentId,
+          componentInstanceId: value.componentInstanceId,
+          name: '', // this.$options.name || '',
+        })
+      })
+
+      value.tree.track(value.onUpdate)
+
+      return state
+    }
+  }) as any
+}
+
+export interface ActionsHook<Context extends IContext<{ actions: {} }>> {
+  (): Ref<Context['actions']>
+  <CB extends (actions: Context['actions']) => object>(
+    cb: CB
+  ): CB extends (actions: Context['actions']) => infer O
+    ? O extends object
+      ? Ref<O>
+      : never
+    : never
+}
+
+export function createActionsHook<
+  Context extends IContext<{ actions: {} }>,
+>(): ActionsHook<Context> {
+  return ((cb?: any): Context['actions'] => {
+    const overmindInstance = inject<any>('overmind')
+
+    return cb ? cb(overmindInstance.actions) : overmindInstance.actions
+  }) as any
+}
+
+export interface EffectsHook<Context extends IContext<{ effects: {} }>> {
+  (): Ref<Context['effects']>
+  <CB extends (effects: Context['effects']) => object>(
+    cb: CB
+  ): CB extends (effects: Context['effects']) => infer O
+    ? O extends object
+      ? Ref<O>
+      : never
+    : never
+}
+
+export function createEffectsHook<
+  Context extends IContext<{ effects: {} }>,
+>(): EffectsHook<Context> {
+  return ((cb?: any): Context['effects'] => {
+    const overmindInstance = inject<any>('overmind')
+
+    return cb ? cb(overmindInstance.effects) : overmindInstance.effects
+  }) as any
+}
+
+export function createReactionHook<Context extends IContext<{ state: {} }>>() {
+  return (): Context['reaction'] => {
+    const overmindInstance = inject<any>('overmind')
+
+    return overmindInstance.reaction
   }
 }
 
-export const createPlugin = (overmind: Overmind<any>) => ({
-  install(
-    Vue: any,
-    propsCallback = ({ state, actions, effects }: any) => ({
-      state,
-      actions,
-      effects,
-    })
-  ) {
-    Vue.mixin(createMixin(overmind, propsCallback))
-  },
-})
-
-export function createConnect<Context extends IContext<any>>(
-  overmind: Overmind<any>
-) {
-  return <T>(
-    cb: ((context: Context) => T) | AnyComponent,
-    component?: AnyComponent
-  ) => {
-    const options: any = component || cb
-    const propsCallback = component ? cb : null
-
-    if (propsCallback && typeof propsCallback !== 'function') {
-      throw new Error(
-        `OVERMIND-VUE: When passing two arguments to "connect", the first has to be a function. You can alternatively only pass a single argument, which is the component`
-      )
-    }
-
-    return {
-      ...options,
-      mixins: (options.mixins ? options.mixins : []).concat(
-        createMixin(overmind, propsCallback, true)
-      ),
-      overmind,
-    } as any
+export function createHooks<
+  Context extends IContext<{ state: {}; actions: {}; effects: {} }>,
+>() {
+  return {
+    state: createStateHook<Context>(),
+    actions: createActionsHook<Context>(),
+    effects: createEffectsHook<Context>(),
+    reaction: createReactionHook<Context>(),
   }
 }
